+++ lib/settings/settings.dart
@@ -1,6 +1,6 @@
 import 'package:Pilll/database/database.dart';
-import 'package:Pilll/components/pill_sheet_type_select_page.dart';
-import 'package:Pilll/components/setting_menstruation_page.dart';
+import 'package:Pilll/components/organisms/pill/pill_sheet_type_select_page.dart';
+import 'package:Pilll/components/organisms/pill/setting/setting_menstruation_page.dart';
 import 'package:Pilll/entity/pill_mark_type.dart';
 import 'package:Pilll/entity/pill_sheet.dart';
 import 'package:Pilll/entity/pill_sheet_type.dart';
